requirements.txt
beautifulsoup4==4.9.0
click==7.1.2
fake-useragent==0.1.11
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.4
numpydoc==0.9.2
pandas==1.0.3
pandas-profiling==2.8.0
patsy==0.5.1
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
regex==2020.5.14
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
simplejson==3.17.0
six==1.14.0
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
tangled-up-in-unicode==0.0.6
Twisted==19.10.0
urllib3==1.25.9
w3lib==1.21.0
